turned CharResponseWrapper into real class
