+++ utilities/tools.descartes.teastore.kieker/src/main/java/tools/descartes/teastore/registryclient/rest/TrackingFilter.java
@@ -133,38 +133,4 @@ public class TrackingFilter implements Filter {
 		CF_REGISTRY.unsetThreadLocalEOI();
 		CF_REGISTRY.unsetThreadLocalESS();
 	}
-	
-	public class CharResponseWrapper extends HttpServletResponseWrapper {
-	    private CharArrayWriter output;
-
-	    public String toString() {
-	        return output.toString();
-	    }
-
-	    public CharResponseWrapper(HttpServletResponse response) {
-	        super(response);
-	        output = new CharArrayWriter();
-	    }
-
-	    public PrintWriter getWriter() {
-	        return new PrintWriter(output);
-	    }
-	    
-	    @Override
-	    public ServletOutputStream getOutputStream() throws IOException {
-	        //This is the magic to prevent closing stream, create a "virtual" stream that does nothing..
-	        return new ServletOutputStream() {
-	            @Override
-	            public void write(int b) throws IOException {
-	            	output.write(b);
-	            }
-	            @Override
-	            public void setWriteListener(WriteListener writeListener) {}
-	            @Override
-	            public boolean isReady() {
-	                return true;
-	            }
-	        };
-	    }
-	}
 }
